+++ CHANGELOG.md
@@ -4,12 +4,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
-## v2.3.2 - 03/06/2026
+## v2.4.0 - 03/09/2026
 
 ### Added
 
 - Allow reductions (min, max, mean, etc.) to be performed over `axis=None`, which will return a single value as a `ComputeMap` object.
 - Allow a custom `earthdaiy.eathone.auth.Auth` instance to be passed into `.compute` and `.visualize` calls, instead of relying on the default auth of the environment.
+- Added a `predicate_filter` keyword argument that allows the user to pass in a filter at the time of construction of a `Mosaic`, instead of filtering later, allowing for better perfomance and less data being pulled from platform.
+- Added a `sort_by` keyword argument that allows the user to pass in an argument for how to sort the data at the time of construction of a `Mosaic`, and an additional optional `ascending` keyword argument for how to sort the data.
 
 ### Changed